Copyright notices & move
- Moved categories into their own package.
- Copyright notice on all files.
