+++ build.gradle
@@ -86,7 +86,6 @@ repositories {
         url 'https://maven.blamejared.com'
     }
     maven {
-        // location of the maven that hosts JEI files
         name = "Progwml6 maven"
         url = "https://dvs1.progwml6.com/files/maven/"
     }
